requirements.txt
Django==1.9.5
django-appconf==1.0.2
django-imagekit==3.3
django-robots==2.0
django-static-sitemaps==4.2.0
django-summernote==0.8.6
gevent==1.1.1
greenlet==0.4.9
image==1.5.4
mysqlclient==1.3.7
pilkit==1.1.13
Pillow==3.3.1
pycparser==2.14
six==1.10.0
